+++ pkg/datasource/resource_handler.go
@@ -2,7 +2,7 @@ package datasource
 
 import (
 	"encoding/json"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"strings"
 
@@ -16,7 +16,7 @@ import (
 // mux.HandleFunc("/", ds.RootHandler)
 // mux.HandleFunc("/strava-api", ds.StravaAPIHandler)
 
-func (ds *StravaDatasource) RootHandler(rw http.ResponseWriter, req *http.Request) {
+func (ds *StravaDatasourcePlugin) RootHandler(rw http.ResponseWriter, req *http.Request) {
 	ds.logger.Debug("Received resource call", "url", req.URL.String(), "method", req.Method)
 
 	_, err := rw.Write([]byte("Hello from Strava data source!"))
@@ -26,12 +26,12 @@ func (ds *StravaDatasource) RootHandler(rw http.ResponseWriter, req *http.Reques
 	rw.WriteHeader(http.StatusOK)
 }
 
-func (ds *StravaDatasource) StravaAuthHandler(rw http.ResponseWriter, req *http.Request) {
+func (ds *StravaDatasourcePlugin) StravaAuthHandler(rw http.ResponseWriter, req *http.Request) {
 	if req.Method != http.MethodPost {
 		return
 	}
 
-	body, err := ioutil.ReadAll(req.Body)
+	body, err := io.ReadAll(req.Body)
 	defer req.Body.Close()
 	if err != nil || len(body) == 0 {
 		writeError(rw, http.StatusBadRequest, err)
@@ -47,7 +47,7 @@ func (ds *StravaDatasource) StravaAuthHandler(rw http.ResponseWriter, req *http.
 	}
 
 	pluginCxt := httpadapter.PluginConfigFromContext(req.Context())
-	dsInstance, err := ds.getDSInstance(pluginCxt)
+	dsInstance, err := ds.getDSInstance(req.Context(), pluginCxt)
 	if err != nil {
 		ds.logger.Error("Error loading datasource", "error", err)
 		writeError(rw, http.StatusInternalServerError, err)
@@ -64,9 +64,9 @@ func (ds *StravaDatasource) StravaAuthHandler(rw http.ResponseWriter, req *http.
 	writeAuthResponse(rw, result)
 }
 
-func (ds *StravaDatasource) ResetAccessTokenHandler(rw http.ResponseWriter, req *http.Request) {
+func (ds *StravaDatasourcePlugin) ResetAccessTokenHandler(rw http.ResponseWriter, req *http.Request) {
 	pluginCxt := httpadapter.PluginConfigFromContext(req.Context())
-	dsInstance, err := ds.getDSInstance(pluginCxt)
+	dsInstance, err := ds.getDSInstance(req.Context(), pluginCxt)
 	if err != nil {
 		ds.logger.Error("Error loading datasource", "error", err)
 		writeError(rw, http.StatusInternalServerError, err)
@@ -96,9 +96,9 @@ func (ds *StravaDatasource) ResetAccessTokenHandler(rw http.ResponseWriter, req 
 	}
 }
 
-func (ds *StravaDatasource) ResetCacheHandler(rw http.ResponseWriter, req *http.Request) {
+func (ds *StravaDatasourcePlugin) ResetCacheHandler(rw http.ResponseWriter, req *http.Request) {
 	pluginCxt := httpadapter.PluginConfigFromContext(req.Context())
-	dsInstance, err := ds.getDSInstance(pluginCxt)
+	dsInstance, err := ds.getDSInstance(req.Context(), pluginCxt)
 	if err != nil {
 		ds.logger.Error("Error loading datasource", "error", err)
 		writeError(rw, http.StatusInternalServerError, err)
@@ -123,12 +123,12 @@ func (ds *StravaDatasource) ResetCacheHandler(rw http.ResponseWriter, req *http.
 	}
 }
 
-func (ds *StravaDatasource) StravaAPIHandler(rw http.ResponseWriter, req *http.Request) {
+func (ds *StravaDatasourcePlugin) StravaAPIHandler(rw http.ResponseWriter, req *http.Request) {
 	if req.Method != http.MethodPost {
 		return
 	}
 
-	body, err := ioutil.ReadAll(req.Body)
+	body, err := io.ReadAll(req.Body)
 	defer req.Body.Close()
 	if err != nil || len(body) == 0 {
 		writeError(rw, http.StatusBadRequest, err)
@@ -144,7 +144,7 @@ func (ds *StravaDatasource) StravaAPIHandler(rw http.ResponseWriter, req *http.R
 	}
 
 	pluginCxt := httpadapter.PluginConfigFromContext(req.Context())
-	dsInstance, err := ds.getDSInstance(pluginCxt)
+	dsInstance, err := ds.getDSInstance(req.Context(), pluginCxt)
 	if err != nil {
 		ds.logger.Error("Error loading datasource", "error", err)
 		writeError(rw, http.StatusInternalServerError, err)